HM #22 - Fixed Night Hound avatars
