+++ src/HoNAvatarManager.Core/Parsers/Ability/AbilityFilesEntityParser.cs
@@ -1,8 +1,6 @@
-﻿using System;
-using System.IO;
+﻿using System.IO;
 using System.Linq;
 using HoNAvatarManager.Core.Extensions;
-using HoNAvatarManager.Core.Helpers;
 
 namespace HoNAvatarManager.Core.Parsers.Ability
 {
@@ -38,62 +36,6 @@ namespace HoNAvatarManager.Core.Parsers.Ability
                     avatarAbilityFile.CopyTo(newAvatarAbilityFilePath.FullName, true);
                 }
             }
-        }
-
-        private string GetAvatarDirectory(string heroDirectoryPath, string avatarKey)
-        {
-            var heroDirectories = Directory.EnumerateDirectories(heroDirectoryPath).Select(d => new DirectoryInfo(d));
-            var avatarDirectory = heroDirectories.FirstOrDefault(d => string.Equals(d.Name, avatarKey, StringComparison.InvariantCultureIgnoreCase));
-
-            if (avatarDirectory != null)
-            {
-                return avatarDirectory.FullName;
-            }
-
-            var parsedAvatarKey = avatarKey.ParseAvatarKey();
-            avatarDirectory = heroDirectories.FirstOrDefault(d => string.Equals(d.Name, parsedAvatarKey, StringComparison.InvariantCultureIgnoreCase));
-
-            if (avatarDirectory != null)
-            {
-                return avatarDirectory.FullName;
-            }
-
-            var entityFilePath = GetHeroEntityPath(heroDirectoryPath);
-            var entityXml = _xmlManager.GetXmlDocument(entityFilePath);
-
-            var entityElement = entityXml.QuerySelector("hero");
-            var entityAvatarElements = entityElement.QuerySelectorAll("altavatar");
-            var entityAvatarElement = entityAvatarElements.FirstOrDefault(a => a.HasKey(avatarKey));
-
-            var modelDirectory = entityAvatarElement.GetAttribute("model").Split("/").First();
-            var avatarDirectoryPath = Path.Combine(heroDirectoryPath, modelDirectory);
-
-            if (Directory.Exists(avatarDirectoryPath))
-            {
-                return avatarDirectoryPath;
-            }
-
-            throw ThrowHelper.DirectoryNotFoundException($"Directory not found for avatar {avatarKey}.");
-        }
-
-        protected string GetHeroEntityPath(string heroDirectoryPath)
-        {
-            var heroEntityFilePath = Path.Combine(heroDirectoryPath, $"hero.entity");
-
-            if (File.Exists(heroEntityFilePath))
-            {
-                return heroEntityFilePath;
-            }
-
-            var heroDirectoryPathInfo = new DirectoryInfo(heroDirectoryPath);
-            var heroNameEntityFilePath = Path.Combine(heroDirectoryPath, $"{heroDirectoryPathInfo.Name}.entity");
-
-            if (File.Exists(heroNameEntityFilePath))
-            {
-                return heroNameEntityFilePath;
-            }
-
-            throw ThrowHelper.FileNotFoundException("Hero entity file not found.", heroNameEntityFilePath);
         }
     }
 }
