+++ src/HoNAvatarManager.Core/Resources/avatar_names_mapping.json
@@ -3278,59 +3278,59 @@
     "avatarInfo": [
       {
         "avatarName": "Teen Hound",
-        "resourceName": ""
+        "resourceName": "Hero_Hantumon.Alt"
       },
       {
         "avatarName": "Night Hunter",
-        "resourceName": ""
+        "resourceName": "Hero_Hantumon.Alt2"
       },
       {
         "avatarName": "Night Hoodlum",
-        "resourceName": ""
+        "resourceName": "Hero_Hantumon.Alt3"
       },
       {
         "avatarName": "Throwback Night Hound",
-        "resourceName": ""
+        "resourceName": "Hero_Hantumon.Classic"
       },
       {
         "avatarName": "Cowardly Night Hound",
-        "resourceName": ""
+        "resourceName": "Hero_Hantumon.Alt4"
       },
       {
         "avatarName": "POG Night Hound",
-        "resourceName": ""
+        "resourceName": "Hero_Hantumon.pog_skin"
       },
       {
         "avatarName": "Rabid Night Bunny",
-        "resourceName": ""
+        "resourceName": "Hero_Hantumon.Alt5"
       },
       {
         "avatarName": "Leo",
-        "resourceName": ""
+        "resourceName": "Hero_Hantumon.Alt6"
       },
       {
         "avatarName": "Fright Hound",
-        "resourceName": ""
+        "resourceName": "Hero_Hantumon.Alt7"
       },
       {
         "avatarName": "Ferocious Leo",
-        "resourceName": ""
+        "resourceName": "Hero_Hantumon.Alt9"
       },
       {
         "avatarName": "Catnip Night Hound",
-        "resourceName": ""
+        "resourceName": "Hero_Hantumon.Alt10"
       },
       {
         "avatarName": "Plushie Night Hound",
-        "resourceName": ""
+        "resourceName": "Hero_Hantumon.Alt11"
       },
       {
         "avatarName": "Misfit Night Hound",
-        "resourceName": ""
+        "resourceName": "Hero_Hantumon.Alt12"
       },
       {
         "avatarName": "Wildcat",
-        "resourceName": ""
+        "resourceName": "Hero_Hantumon.Alt13"
       }
     ]
   },
